requirements.txt
absl-py==0.7.0
astor==0.7.1
certifi==2019.3.9
Click==7.0
Flask==1.0.2
gast==0.2.2
grpcio==1.19.0
h5py==2.9.0
itsdangerous==1.1.0
Jinja2==2.10
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
Markdown==3.0.1
MarkupSafe==1.1.1
mock==2.0.0
numpy==1.16.2
pbr==5.1.3
protobuf==3.7.0
PyYAML==3.13
scipy==1.2.1
six==1.12.0
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
waitress==1.2.1
Werkzeug==0.14.1
